Refactored scripts to 'loess' folder.
