--- local_regression/__init__.py
@@ -1,2 +1,0 @@
-"""Initialize the local_regression package."""
-from .loess import Loess
