+++ loess/__init__.py
@@ -1,0 +1,2 @@
+"""Initialize the loess package."""
+from .loess import Loess
